updated return responses

+++ src/main/java/com/sermon/mynote/repository/UserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import com.sermon.mynote.domain.User;
 
 
-public interface UserRepository extends PagingAndSortingRepository<User, Long>{
+public interface UserRepository extends PagingAndSortingRepository<User, Integer>{
 
 }
